django/requirements.txt
appdirs==1.4.0
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.3
gunicorn==19.6.0
httpie==0.9.9
Markdown==2.6.8
packaging==16.8
Pygments==2.2.0
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
psycopg2==2.6
